Fixed an embedded test failure, and a flaw in STOP-SLAVE parallel
The test is guarded to not run on embedded.

In a case of STOP SLAVE that separates SA (has been done) and CA
(will not 'cos of the stop), see rpl_start_alter_restart_slave.test,
an assert got fired when a "inadvertently" commented-out statement has
been finally identified and restored for action.
It turns out the split alter executing parallel workers did not fully comply
with the stop slave's worker stop protocol and could be on loose
at time the SQL thread thought their have been docked.
It's fixed (and somewhat simplified) with leveraging a "direct"
execution mode (by CA).
SA and the shutdowner threads continue communicating as before
but the latter sets the direct mode ON for a possible "unlikely" CA.
SA may also find its shutdown status at time of changing the slave
gtid state.

In terms of use case the slave stop schema works as the following.
Let's consider sequences of execution (comma separated below) events

1. SA, STOP-SLAVE, *no-more-events*
   interrupts SA after its change to the (gtid) slave state.
   ALTER query will be re-run by a matching CA (if it will show up).

2. SA, STOP-SLAVE, CA
   The same as above and CA indeed re-runs the whole ALTER
   in "direct" mode.

3. *no* SA, STOP-SLAVE, CA
   The same as above in the part of re-running CA (SA might have
   been tried to reach its wait-for-master decision state in the past).