MDEV-27349 ASAN use-after-poison in Query_log_event
The poisoning comaplaint was caused by that
Rollback ALTER does not assign anything to its THD::lex::sql_command,
so it found the value from a previous event and 'cos it was not
ALTER, an inmpropriate branch of ctor was attempted to be
executed.

Fixed with forcing THD::lex::sql_command to denote ALTER indeed.