This commit converts an assert to an error
The assert was caused by unexpected by the slave applier
appearance of a temporary table that shadowed the regular one
for START ALTER phase.

This situation however is not really a replication fault and can
occur practically 'cos of misconfiguration that would made
the master and slave table types diverge in the temporary attribute
sense.

Also a warning is added to catch cleared errors of bin-logged queries
such as CREATE-or-REPLACE.