Corrects the replication alter state in "special" goto-end_inplace branches
Two mysql_alter_table execution branches missed to invoke
write_bin_log_start_alter and wait_for_master prior to process_master_state.

This is corrected to indentify the need and  call them inside
process_master_state.
As these branches are fast thanks to the goto, so calling the two
functions at the receiving end of the goto is fine for performance
as well.